Extracted out CapiSingle model per Regis' comments
